sowsystem/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sowsystem"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "aws-opentelemetry-distro >= 0.10.0",
    "bedrock-agentcore >= 1.3.1",
    "bedrock-agentcore-starter-toolkit >= 0.2.10",
    "boto3 >= 1.34.0",
    "ddtrace >= 2.10.0",
    "mcp >= 1.19.0",
    "opentelemetry-api >= 1.20.0",
    "opentelemetry-sdk >= 1.20.0",
    "pydantic >= 2.0.0",
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
    "python-dotenv >= 1.0.0",
    "strands-agents >= 1.25.0",
    "strands-agents-tools[use_browser,mem0_memory] >= 0.2.20"
]